metrics now recorded for all plugins and labeled in prometheus
